NotificationArea: Add more missing features in GtkNotificationAreaBox
With gtk-sharp 3.0, we can now implement all the features of our
notification area using Gtk.StatusIcon, in particular the custom
tooltip. This commit complements what was already done by Olivier
earlier on.

This might not work perfectly yet, some further testing is needed
when Banshee is running properly on GTK 3.