build: Default to our internal copy of notify-sharp
Bring back the option to use the system-installed notify-sharp, but it
will not be the case by default. As notify-sharp upstream depends on
gtk-sharp-2.0 and hasn't seen a release for quite a while, we now
default on using to our internal copy.

You can still force the use of the system-installed notify-sharp by
adding the --with-system-notify-sharp configure flag. But do that at
your own risk, it will probably cause a build failure.