StreamPositionLabel: Fix rendering of the label and the progress bar
Render the label in the right position, and use the appropriate classes
to have a good-looking progress bar.

It is only now that I was able to understand the original trickery in
drawing that widget. If the progress bar is visible, we first draw the
part of the text that is over the progress bar, to have it in the right
color.