NotificationArea: Add support for GNOME Shell's notification scheme
Use the new features added in the latest version of the Desktop
Notifications Specification: persistent notifications, image-data hint,
and action icons.

As a consequence, we now create a status icon only if the notification
daemon doesn't support persistent notifications (or if it's not there at
all).

As those are all optional features, this should still work with older
notification daemons. But we now always send paths to unscaled images,
so the daemon should be able to handle scaling itself.