Refactor our SearchEntry widget to use new features of Gtk.Entry
The standard GTK.Entry widget now supports primary and secondary icons,
so we don't need to handle the search and clear icons by ourselves.

SearchEntry is now derived directly from Gtk.Entry, and much simpler
than the old composite widget. It only adds a few tweaks to match the
behavior we want, and handles displaying the EmptyMessage. This also
allows use to drop HoverImageButton.

As a non-trivial side effect, this fixes the appearance of the
widget with GTK+ 3.

Adapt the various places using SearchEntry, simplifying where we can.