build: GSettingsSchemaExtractor now inspects all assemblies at once
A <schemalist> tag can contain various <schema> elements. There's not clear
benefit that I can see for now for separating all them in different files,
so from now on it will join them all together in a schema file with the name
'org.gnome.banshee.gschema.xml'.

This will also simplify the generation of the schemas in the makefiles.

Added also some logging to know the progress and if the tool succeeded.