Lastfm: simplify previous tests addition by dropping an interface
The interface IWebRequestCreate already exists in .NET class libraries,
so use it in favour of our own IWebRequestCreator. Only difference is
that this one uses System.Uri instead of string as the type of the uri
parameter, so we need a bit of refactoring to make this compile.

Tests still pass of course. No change of behaviour in this commit.