GSettings: add SchemaEntry properties for FilteredSourceContents children
The configuration settings for the position of the pane containers
implemented as derived classes of the FilteredSourceContents class
(CompositeTrackSourceContents, InternetRadioSourceContents and
PodcastSourceContents) now receive two SchemaEntry objects for the
position of their container, rather than treating them as the special
case of the filter "-1" when constructing the namespace to retrieve its
configuration.

This is a bit more GSettings friendly, but there's more work left to do
in this class afterwards (for the positive filters).