build: refactoring to remove dupe $DLL_MAP_VERIFIER_ASSEMBLY target
After the fix for bgo#691696 landed, the target $DLL_MAP_VERIFIER_ASSEMBLY
is defined in two places. This commits avoids this, by making the Makefile
from the build/ dir include the generic build.rules.mk like other Makefiles
in the build already do.

The only change in this commit that seems to not be related is the move of
the 'if ENABLE_TESTS' block. The reason for it was to avoid this build
failure which would happen if it's not moved:

  Running automake --gnu --add-missing --force --copy -Wno-portability -Wno-portability ...
  build/build.rules.mk:11: LINK must be set with `=' before using `+='

No change of behaviour in this commit.