Configuration: remove static helper methods from ConfigurationClient
If someone needs to do some refactoring around IConfigurationClient it is
very painful because there are a lot of duplicated methods: the static ones
in class ConfigurationClient and the extension methods in the file
IConfigurationClient.cs. So, with this commit we remove the static methods
to unify the API a bit. Instead, we now use the ConfigurationClient
singleton instance everywhere.

To avoid the redundancy of calling this "ConfigurationClient.Client", we
now rename the singleton to be named "Instance", as it is done in other
parts of the code with other singleton instances.

No change of behaviour in this commit.