Lastfm: avoid unnecessary wait when submitting successful scrobbles
In the case where everything was successful but there are still entries in
the queue to submit, the overall connection state was set to 'Idle'.

This meant that on the next interval we would realize that we had tracks to
submit and then set the state to 'NeedTransmit'. This was an unnecessary
interval step and made Banshee wait a bit before attempting to submit the
rest of the queue.

Signed-off-by: Andres G. Aragoneses <knocte@gmail.com>