GSettings: extractor to find internal SchemaEntry instances too
There were some internal/private SchemaEntry instances in some classes
that the extractor should be able to inspect.

For this, a new unit test is added, and the following issue is fixed:
when finding these new set of SchemaEntry instances in the AlbumDialog
class in Muinshe, a NullReferenceException [1] would be thrown because
the access to the class via Reflection would trigger the execution of
its static ctor, which would depend on Banshee running properly to work
(as it would inspect MusicLibrarySource.CurrentFilters property). To fix
this, the static ctor is removed, and the usage of the album_model field
(that was being initialized in this static ctor) is replaced by a property
which is lazily initialized.

[1] Unhandled Exception:
System.TypeInitializationException: An exception was thrown by the type initializer for Muinshee.AlbumDialog ---> System.NullReferenceException: Object reference not set to an instance of an object
  at Muinshee.BaseDialog.get_Music () [0x00000] in /home/andres1210/Code/bansheeGTK3/src/Clients/Muinshee/Muinshee/BaseDialog.cs:154
  at Muinshee.AlbumDialog..cctor () [0x00000] in <filename unknown>:0
  --- End of inner exception stack trace ---
  at (wrapper managed-to-native) System.Reflection.MonoField:GetValueInternal (System.Reflection.MonoField,object)
  at System.Reflection.MonoField.GetValue (System.Object obj) [0x0006a] in /home/andres1210/Code/system_mono/mono/mcs/class/corlib/System.Reflection/MonoField.cs:125
  at GSettingsSchemaExtractorProgram+GSettingsKey.ExtractFromField (SchemaSet schemaSet, System.Reflection.FieldInfo field) [0x00000] in <filename unknown>:0
  at GSettingsSchemaExtractorProgram.Extract (IEnumerable`1 types) [0x00000] in <filename unknown>:0
  at GSettingsSchemaExtractorProgram.Extract (System.IO.DirectoryInfo dir) [0x00000] in <filename unknown>:0
  at GSettingsSchemaExtractorProgram.Main (System.String[] args) [0x00000] in <filename unknown>:0