SaveTrackMetadataJob: check if PathPattern is null before renaming file
Since this change[1] in DatabaseTrackInfo, PathPattern property may return
null instead of a default PathPattern. Even though RenameFile () was
already checking if the track belonged to the MusicLibrarySource, this
check for null, to return early, is faster, and safer (because in the
future someone could implement RenameFile operation in non-Music sources
and could forget to do this null check).

[1] https://git.gnome.org/browse/banshee/commit/?id=2c82ffbeaf4f4ff9f996b4dfaf1bfc84064353f5