build: update post-build.bat (autogenerate) and clarify README (windows)
Turns out post-build.bat cannot be edited manually because it is a
generated file, so I'm adding some clarifications about this in the
README.txt file and on the file itself with a REM comment.

I've also auto-generated the file now, which has re-fixed the 2 issues I
fixed recently in it (but in a different way), and has added some copy
lines of more resources I wasn't aware of.