Feed: Avoid loop when podcast has date values in the future (bgo#706173)
CheckForItemsToDownload needs to save the feed, which in turn would call
CheckForItemsToDownload. This could cause an infinite loop if both the
feed LastBuildDate and an item's PubDate were set in the future.

To avoid that, add a new Save method with a parameter to avoid calling
CheckForItemsToDownload, and use that method in CheckForItemsToDownload.

This is not a great API (with 2 bool parameters), but it's the best I
could come up with to avoid breaking existing API or risking unwanted
side-effects.