SourceManager: avoid warnings when shutting down (bgo#709376)
When shutting down banshee while a track is being played, the disposal
of the SourceManager instance was causing a NRE, which was printed as
a Warning in the console. The culprit of the NRE consisted on the class
FileSystemQueue trying to access SourceManager.MusicLibrary property
in the Dispose() method to unsubscribe from some events, but MusicLibrary
property had already been set to null by SourceManager's Dispose() method.

The solution is simply to delay the assignment of null values to these
properties to *after* the SourceManager children sources have been disposed.

This wasn't causing a crash or anything severe, but it's good that we get
rid of the noise that it was generating.