configure: move GTK backend detection from libbanshee to bnpx11 macro
The variables HAVE_X11, HAVE_QUARTZ, GRAPHIC_SUBSYSTEM_X11 and
GRAPHICS_SUBSYSTEM_QUARTZ were not being used. The one which was
being used (in the config summary) but not filled with any value,
was GRAPHICS_SUBSYSTEM.

There was already some GTK backend detection in bnpx11 m4 macro so
we move from libbanshee to this just the only bits needed to fill
the value GRAPHICS_SUBSYSTEM, also because libbanshee will be soon
just optional.