PlayerInterface: kill Gtk CRITICAL in the console at shutdown (bgo#710266)
When pausing a video and closing banshee, a CRITICAL from GTK was
being thrown in the console because Nereid's PlayerInterface was
still subscribed to some Service events even after disposed, which
in turn would cause sending some disposed widget's pointer (NULL)
to Gtk.

The fix is to unsubscribe from this events in the Dispose() method.

This is the second and last fix for BGO#710266.