Rename the MeeGo extension to MediaPanel and drop MeeGo integration
Rename everything from MeeGo to MediaPanel, with the appropriate
adjustments. All the MeeGo integration is also removed, so we can build
this extension by default. But the extension is not active by default.

The general idea for the MediaPanel extension is to provide a simple UI
to quickly browse and play your media. A bit like the MiniMode, but
less minimal. It is not specifically targetted  for netbooks, but it
will probably be more interesting for smaller screens, and maybe touch
devices ?