Backends: drop the Banshee.NowPlaying.X11 extension
This extension was causing a bug when using the Fullscreen
functionality at video playback. It may have been useful
in the old gtk2 days, but today fullscreen works well without
it (tested with Ubuntu and Gnome3).

This also gets us closer to the goal of having 0% unmanaged
code in banshee.