MediaPanelContents: Put the artist and album ListView in a Paned
The ListView widget doesn't have a Gdk.Window of its own, it uses the
one of its parent. So, it seems that if 2 ListViews are put in the same
Gtk.Box, they end up sharing the same Gdk.Window. This has some unwanted
consequences, like mouse events not being handled at the right
coordinates. See bgo#710547.

So we put the artist and album ListView widgets in a Paned instead of a
Box, which gets us the right behavior. In addition, it make the division
between the two ListView adjustable, which is desirable in its own
right.