MediaPanel: Make the extension enabled by default
Currently, if the MediaPanel extension is disabled, running the
MediaPanel client will just silently fallback to the main UI. To avoid
this, the extension is now enabled by default, until we find some better
way to do this.

Furthermore, the extension is intentionaly not put in a category, so it
won't appear in the list of extensions in the preferences dialog, thus
can't be disabled.