libbanshee: remove leftover from the port to Gst1.0 (stream-changed)
When porting this backend to GStreamer 1.0 [1], a new switch case
was introduced for the GST_MESSAGE_STREAM_START message type, but
removing the old way[2] of catching this event (via another message
of type _ELEMENT with a "playbin2-stream-change" structure name)
was not done. For reference, this change in GStreamer happened in
this commit [3].

This should be considered as part of BGO#679433 too.

[1] https://git.gnome.org/browse/banshee/commit/?id=e40923df1bc55129832dff5ca2c782f5040b412f
[2] https://git.gnome.org/browse/banshee/commit/?id=b1519fded83eab0f0d2dba821dcecf802f54490b
[3] http://cgit.freedesktop.org/gstreamer/gst-plugins-base/commit/?id=a0fbf92d43e81452b6cce09ee20d9a8cc4af7b06