libbanshee: remove obsolete comment about 'about-to-finish' hook
The callback attached to the 'about-to-finish' signal of Playbin served
two purposes some time ago:
a) Requesting the next track URI to banshee managed land.
b) Setting up a timeout to fire the OnNextTrackStarting callback some
moments after the about-to-finish signal was fired.

After this commit [1] from Christopher, the purpose (b) above was not
needed anymore because GStreamer would fire the event itself internally.
This change would render this comment *obsolete* (but Christopher probably
forgot to remove it) and *misleading*, the latter because the comment
only applied to the (b) purpose above, not (a). We still need to request
the next track URI to banshee's managed land, so the 'about-to-finish'
signal still needs to be used (contrary to what the comment seems to
suggest).

NOTE: Christopher's commit[1] hooked to the "playbin2-stream-changed"
event, but this has been deprecated in GStreamer in favour of the msg
type STREAM_START. More info here[2].

[1] https://git.gnome.org/browse/banshee/commit/?id=b1519fded83eab0f0d2dba821dcecf802f54490b
[2] https://git.gnome.org/browse/banshee/commit/?id=ea128c28c2fc167039cfbe1a8ae489c55fa55c0e