MediaPanel: disabling MediaPanel's theme
Since commit[1] happened, whole Banshee feels a bit more ugly and
clunky, and it is because the MediaPanel extension changes the
theme when it's loaded (even if it's not visible yet). This was
clearly not intended.

We should probably change it so it toggles between one theme or the
other when MediaPanel is shown or hidden, but this theme doesn't
look good either when showing the MediaPanel UI in Ubuntu's GTK3.

It's very likely that this theme was very specific to Meego's
rendering, so let's disable it for now, until someone has the
energy to fix it.

[1] https://git.gnome.org/browse/banshee/commit/?id=39dd08cd517300999ed3124691cbf8973887423d