GStreamerSharp: port to GTK3
Even though GStreamerSharp backend doesn't strictly depend on GTK at
compile time, it uses GDK at runtime to get the handle of a Window
in order to tell GStreamer where to render a video. To do this,
it uses a P/Invoke to an unmanaged library which was still pointing
to GTKv2.0 instead of GTKv3.0. This was causing the following errors:

(Nereid:16860): GLib-GObject-WARNING **: cannot register existing type `GdkWindow'

(Nereid:16860): Gdk-WARNING **: /build/buildd/gtk+2.0-2.24.17/gdk/x11/gdkdrawable-x11.c:952 drawable is not a pixmap or window

We leave a couple of FIXMEs for Windows/Mac platforms.

Signed-off-by: Andrés G. Aragoneses <knocte@gmail.com>