GStreamerSharp: make video playback work properly again
VideoManager initialization being commented out meant that video
playback was working, but not in the proper place (a new window
would pop-up instead of playing the video inside the NowPlaying
area). This was caused by the bad usage of the Gst# method
Gst.Bin.GetByInterface(), which receives a GType, but was getting
a Type of an interface that corresponds to the GInterface type.
The way to do it is using the -Adapter class that maps to this
GInterface.

We commented out other lines that cause the player to freeze/hang
altogether (probably related to some threading deadlock). Commenting
them out means that GStreamerSharp still has a bug with regards to
playing a second video while a first one is already being played,
but the latter bug should be easier to solve, for now, than the
freezes.