GStreamerSharp: prepare overlay the soonest possible if it's a video
There was still a small (but noticeable) time window between the start
of a video playback and the mapping of its rendering box to the
NowPlaying window. Doing it just after the Play() command is sent to
engine is the soonest possible it can be done to avoid this visual
glitch.