GStreamerSharp: prevent NRE from happening if there is no next track
If there is no next track (i.e.: you have only 1 video in your library
and you play that video), a NullReferenceException would happen at the
end of the playback of the current track.

This null check is brought as a pristine copy of the one that exists
on the unmanaged backend.