GStreamerSharp: fix playback of subsequent videos
If the user tried to play a 2nd video (requesting it by hand, not just
waiting for the 1st video to end), it would be rendered outside the
NowPlaying area. This was because there was no way to invalidate
previous overlays.

By looking at the code of the unmanaged Banshee.GStreamer backend, I've
noticed that:

 - In it, the invalidation happens more frequently (every time the
 function bp_video_find_video_overlay() is called.
 - The invalidation happens because of two possible events that call
 the former function:
   * bp_video_sink_element_added
   * bp_video_bus_element_sync_message

The `element_added` and `sync_message` hooks are precisely the ones
that are causing hangs in Banshee using the new managed GStreamerSharp
backend, but since [1] I've managed to make video working without
them (thanks to the Expose() call, but improving it later in [2]).

Now, after this commit, video is completely fixed 100%, and by a
means that doesn't happen so often as in the unmanaged backend, which
means that this backend would have better performance, plus we get rid
of unnecessary hooks (less code, less problems). (Anyway, we leave them
commented for now, I'm more comfortable removing them a bit later.)

Now we should be all set for the next phase: migration to Gst# 1.0
bindings, which target GStreamer >= 1.0. (Something which is already
work in progress.)

[1] https://git.gnome.org/browse/banshee/commit/?id=96ee69d2534f12dd40aabd4e53ada7337a898f41
[2] https://git.gnome.org/browse/banshee/commit/?id=4566e800b4b1402491251cba164f8deef20113f9